Set certname to dns name
Previous to this commit, autoami used a trick of appending the group
name to the certificate name of the agent.  This was extremely hacky and
required special node definitions in the manifest.  This approach uses
the VCS hook to save the data to a yaml file on the system. The report
processor loads this file to see if the reporting node is listed in the
yaml file.  This way the node can be classified using the --node-group
parameter built in to cloud provisioner.  Unforunately, this means the
VCS hook must be run on the puppet master. A central database should be
used in the future.